'版本v1'

+++ Blueprint/scheduler_task.py
@@ -8,27 +8,118 @@ from modules.Tables import *
 from sqlalchemy import and_,func,update
 from flask_jwt_extended import create_access_token, get_jwt_identity, jwt_required
 from Video_Quality_Diagnosis.VQD import *
+from sqlalchemy.exc import SQLAlchemyError
+
 
 # 创建蓝图
 scheduler_task = Blueprint('scheduler_task', __name__)
+
+# 质量诊断数据写入
+def perform_diagnostic_task(data):
+    try:
+        # 开始事务
+        with db.session.begin_nested():
+
+            # 查找并批量移动现有数据到诊断历史表
+            existing_data = db.session.query(Diagnosis_data).all()
+
+            if existing_data:
+                # 创建要插入到历史表中的数据列表
+                history_entries = [
+                    Diagnosis_data_old(
+                        equipment_id=entry.equipment_id,
+                        equipment_ip=entry.equipment_ip,
+                        equipment_name=entry.equipment_name,
+                        vcr_ip=entry.vcr_ip,
+                        diagnosis_type=entry.diagnosis_type,
+                        db101=entry.db101, db102=entry.db102, db103=entry.db103,
+                        db104=entry.db104, db105=entry.db105, db106=entry.db106,
+                        db107=entry.db107, db108=entry.db108, db109=entry.db109,
+                        db110=entry.db110, db111=entry.db111, db112=entry.db112,
+                        db113=entry.db113, db114=entry.db114, db115=entry.db115,
+                        create_time=entry.create_time
+                    )
+                    for entry in existing_data
+                ]
+
+                # 批量保存到历史表
+                db.session.bulk_save_objects(history_entries)
+
+            # 清空现有的诊断数据表
+            db.session.query(Diagnosis_data).delete()
+
+            # 准备批量插入的新诊断数据
+            new_entries = [
+                Diagnosis_data(
+                    equipment_id=i[0], equipment_ip=i[1], equipment_name=i[2], vcr_ip=i[3],
+                    diagnosis_type=i[4], db101=i[5], db102=i[6], db103=i[7],
+                    db104=i[8], db105=i[9], db106=i[10], db107=i[11],
+                    db108=i[12], db109=i[13], db110=i[14], db111=i[15],
+                    db112=i[16], db113=i[17], db114=i[18], db115=i[19],
+                    create_time=datetime.now().strftime("%Y-%m-%d %H:%M:%S")
+                )
+                for i in data
+            ]
+            # 批量插入新的诊断数据
+            db.session.bulk_save_objects(new_entries)
+
+        # 提交事务
+        db.session.commit()
+        print('诊断数据上传完成！')
+
+    except SQLAlchemyError as e:
+        # 回滚事务
+        db.session.rollback()
+        print(f'发生错误: {str(e)}')
+
+# 质量诊断数据写入
+# def perform_diagnostic_task(data):
+    #
+    # # 查找诊断数据表，将诊断数据表数据写入诊断历史表
+    # db.session.query(Diagnosis_data).all()
+    # # 清空诊断数据表
+    #
+    # # 数据插入诊断数据表
+    # for i in data:
+    #     res = Diagnosis_data(
+    #         equipment_id=i[0],equipment_ip=i[1],equipment_name=i[2],vcr_ip=i[3],
+    #         diagnosis_type=i[4],db101=i[5],db102=i[6],db103=i[7],
+    #         db104=i[8],db105=i[9],db106=i[10],db107=i[11],
+    #         db108=i[12],db109=i[13],db110=i[14],db111=i[15],
+    #         db112=i[16],db113=i[17],db114=i[18],db115=i[19],
+    #         create_time=datetime.now().strftime("%Y-%m-%d %H:%M:%S")
+    #     )
+    #     db.session.add(res)
+    #     db.session.commit()
+    # print('诊断数据上传完成！')
 
 
 
 # 质量诊断运行（定时任务设置）
 def run_task(task_id,task_name,task_interval_seconds,time_type,rtsp_list,diagnosis_type_list,scheduled_concent,scheduled_status):
-    print('当前任务id为:',task_id,'当前任务名称为:',task_name,'当前时间为:',task_interval_seconds,'时间类型:',time_type,
-          'rtsp:',rtsp_list,'类型:',diagnosis_type_list,'备注：',scheduled_concent,'状态: ',scheduled_status)
+    print('----------------------------------------------------------------------------------------')
+    details = [
+        f"当前任务id为: {task_id}",
+        f"当前任务名称为: {task_name}",
+        f"当前时间为: {task_interval_seconds}",
+        f"时间类型: {time_type}",
+        f"rtsp: {rtsp_list}",
+        f"类型: {diagnosis_type_list}",
+        f"备注: {scheduled_concent}",
+        f"状态: {scheduled_status}"
+    ]
+    print("\n".join(details))
     try:
         print(f"运行任务: {task_name} 于 {datetime.now()}")
-        # data1 = [{'id': '1', 'name': '监控点1', 'rtsp': 'rtsp://admin:Pc@12138@192.168.7.77'},
-        #          {'id': '2', 'name': '监控点2', 'rtsp': 'rtsp://admin:Pc@12138@192.168.7.31'}]
-        # data2 = [{'id': '1', 'name': '场景变化', 'prewarn_threshold': 30, 'warn_threshold': 70},
-        #          {'id': '2', 'name': '视频冻结', 'prewarn_threshold': 30, 'warn_threshold': 70}]
-        # res = VQD.run(eval(rtsp_list), eval(diagnosis_type_list))
-        # print(res)
-        print('----------------------------------------------------------------------------------------')
+        # 手动创建应用上下文
+        res = VQD.run(eval(rtsp_list), eval(diagnosis_type_list))
+        print(res, '诊断数据')
+        # 在函数内部导入 app
+        from main import app
         # 调用任务的处理函数
-        # perform_diagnostic_task(task)
+        with app.app_context():
+            perform_diagnostic_task(res)
+        print('----------------------------------------------------------------------------------------')
     except Exception as e:
         print(f"运行任务时发生异常: {e}")
 
@@ -171,6 +262,11 @@ def update_task():
     if not task_data:
         return jsonify({"msg": "任务不存在",'code':404})
 
+    # 检查 scheduled_status 是否为 '启用' 或 '禁用'
+    if scheduled_status != '启用' and scheduled_status != '禁用':
+        return jsonify({"msg": "参数错误", 'code': 404})
+
+
     scheduler = current_app.apscheduler
     job_id = f"task_{task.id}"
 
@@ -262,8 +358,10 @@ def update_task():
         scheduler.pause_job(job_id)
         print('停止当前任务---:',job_id)
 
-
     else:
+
+        if not existing_job:
+            return jsonify({"msg": "任务已更新", 'code': 200})
 
         # 不启用停止，进行修改操作
         if time_type == 'seconds':
@@ -301,20 +399,34 @@ def update_task():
 # 定时任务删除
 @scheduler_task.route('/delete_task', methods=['POST'])
 def delete_task():
-    id = request.form.get('id') # 任务id
+    id = request.form.get('id',None) # 任务id
 
-    task = db.session.query(ScheduledTask).filter(ScheduledTask.id==id).first()
-    task_data = convert_to_dict(task,['id','name','interval_seconds'])
-    if not task_data:
-        return jsonify({"msg": "任务不存在",'code':404})
+    if id:
+        try:
+            result = eval(id)  # 解析
+            if isinstance(result, list):  # 检查是否是一个列表
+                for i in eval(id):
+                    task = db.session.query(ScheduledTask).filter(ScheduledTask.id == i).first()
+                    task_data = convert_to_dict(task, ['id', 'name', 'interval_seconds'])
+                    if not task_data:
+                        return jsonify({"msg": "任务不存在", 'code': 404})
 
-    scheduler = current_app.apscheduler
-    job_id = f"task_{task.id}"
+                    scheduler = current_app.apscheduler
+                    job_id = f"task_{task.id}"
 
-    scheduler.remove_job(job_id)
+                    # 检查任务是否存在
+                    existing_job = scheduler.get_job(job_id)
+                    if existing_job:
+                        scheduler.remove_job(job_id)
 
-    db.session.delete(task)
-    db.session.commit()
+                    db.session.delete(task)
+                    db.session.commit()
+            else:
+                return jsonify({"msg": "id不为所传参数校验标准", 'code': 400})
+        except Exception as e:
+            return jsonify({"msg": "id不为所传参数校验标准", 'code': 400})
+    else:
+        return jsonify({"msg": "id不为所传参数校验标准", 'code': 400})
 
     return jsonify({"msg": "任务已删除",'code':200})
 
@@ -322,6 +434,9 @@ def delete_task():
 # 定时任务展示接口
 @scheduler_task.route('/show_task', methods=['GET'])
 def show_task():
+    # 搜索内容
+    name = request.args.get('name')
+
     # 类型
     time_type = request.args.get('time_type')
     # 状态
@@ -334,6 +449,10 @@ def show_task():
 
     # 假设前端传递的参数为可能有一个或多个
     filters = []
+
+    # 搜索名称
+    if name:
+        filters.append(ScheduledTask.name.ilike(f"%{name}%"))
 
     # 如果存在类型参数，将其加入过滤条件
     if time_type:
@@ -369,7 +488,7 @@ def show_task():
         'rtsp_list': i.rtsp_list,
         'diagnosis_type_list': i.diagnosis_type_list,
         'scheduled_concent': i.scheduled_concent,
-        'scheduled_status': status[int(i.scheduled_status) - 1]['value'],
+        'scheduled_status': vcr_status[int(i.scheduled_status) - 1]['value'],
 
     } for i in res.items]
 
@@ -384,28 +503,31 @@ def show_task():
     return jsonify({'code':200,'msg':'查询成功','data':response_data})
 
 
-# 定时任务状态修改接口
+# 定时任务状态详情接口
 @scheduler_task.route('/task_detail', methods=['GET'])
 def task_detail():
     # 任务id
-    id = request.form.get('id')
+    id = request.args.get('id')
 
     # 查找当前算法仓运行状态id
     data = db.session.query(ScheduledTask).filter(ScheduledTask.id == id).first()
 
-    res_data = {
-        'id': data.id,
-        'name': data.name,
-        'interval_seconds': data.interval_seconds,
-        'next_run_time': data.next_run_time,
-        'create_time': (data.create_time).strftime("%Y-%m-%d %H:%M:%S"),
-        'time_type': data.time_type,
-        'rtsp_list': data.rtsp_list,
-        'diagnosis_type_list': data.diagnosis_type_list,
-        'scheduled_concent': data.scheduled_concent,
-        'scheduled_status': status[int(data.scheduled_status) - 1]['value'],
+    if data:
+        res_data = {
+            'id': data.id,
+            'name': data.name,
+            'interval_seconds': data.interval_seconds,
+            'next_run_time': data.next_run_time,
+            'create_time': (data.create_time).strftime("%Y-%m-%d %H:%M:%S"),
+            'time_type': data.time_type,
+            'rtsp_list': data.rtsp_list,
+            'diagnosis_type_list': data.diagnosis_type_list,
+            'scheduled_concent': data.scheduled_concent,
+            'scheduled_status': vcr_status[int(data.scheduled_status) - 1]['value'],
+        }
+        return jsonify({'code': 200, 'msg': '查询成功', 'data': res_data})
+    else:
+        return jsonify({'code': 400, 'msg': '该数据不存在'})
 
-    }
 
-    return jsonify({'code':200,'msg':'修改成功','data':res_data})
